Add error handling
In the previous version, the website directory was always deleted if there was data in the directory. This is a problem if the quarto render command fails, because in this case no new data is available and the old data is deleted.

In the new version, the shell aborts further execution as soon as a command fails and should ideally display a short message.