Ignore model.qmd and url to model site
Since the dynamic model cannot be converted into a static website, we need a different approach. 

The first idea is to create a permanently running container with the Shiny app and then let the proxy do everything else. 

For this reason, the link (href) was replaced by the (url) and the model.qmd was ignored.